feat: product apis created and tested crud

+++ node-backend/src/controllers/admin/product/product.controller.js
@@ -1,7 +1,8 @@
 import { Product } from "../../../models/product.model.js";
-import { check, validationResult } from "express-validator";
+import { body, check, query, validationResult } from "express-validator";
 import ApiError from "../../../utils/apiErrors.js";
 import ApiResponse from "../../../utils/apiResponse.js";
+import { isValidObjectId } from "../../../utils/mongoose.utility.js";
 
 
 const productValidationRules = [
@@ -23,7 +24,7 @@ const addProduct = async (req, res) => {
     try {
         const existingProduct = await Product.findOne({ name: name, vendorId: vendorId, brandName: brandName });
         if (existingProduct) {
-            return res.status(400).json(new ApiError(400, null,"Product already exists"));
+            return res.status(400).json(new ApiError(400, null, "Product already exists"));
         }
         const newProduct = new Product(req.body);
         await newProduct.save();
@@ -37,27 +38,47 @@ const addProduct = async (req, res) => {
 
 const updateProduct = async (req, res) => {
     const { productId } = req.params;
+    if (!isValidObjectId(productId)) {
+        return res.status(400).json(new ApiError(400, null, "Invalid product ID"));
+    }
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         return res.status(400).json(new ApiError(400, "Validation Error", errors.array()));
     }
 
     try {
-        const updatedProduct = await Product.findByIdAndUpdate(productId, req.body, { new: true });
-        if (!updatedProduct) {
+        const existingProduct = await Product.findById(productId);
+        if (!existingProduct) {
             return res.status(404).json(new ApiError(404, "Product not found"));
         }
+        const updatedData = {
+            ...req.body,
+            specifications: req.body.specifications
+                ? { ...req.body.specifications }
+                : existingProduct.specifications,
+            images: req.body.images
+                ? [...new Set([...existingProduct.images, ...req.body.images])]
+                : existingProduct.images,
+        };
+
+        // Update the product
+        const updatedProduct = await Product.findByIdAndUpdate(productId, updatedData, { new: true });
         return res.status(200).json(new ApiResponse(200, updatedProduct, "Product updated successfully!"));
     } catch (error) {
         console.error(error);
         return res.status(500).json(new ApiResponse(500, "Something went wrong", [error.message]));
     }
 };
+
+
 /*--------------------------------------------Delete a product---------------------------------------*/
 
 // 
 const deleteProduct = async (req, res) => {
     const { productId } = req.params;
+    if (!isValidObjectId(productId)) {
+        return res.status(400).json(new ApiError(400, null, "Invalid product ID"));
+    }
 
     try {
         const deletedProduct = await Product.findByIdAndDelete(productId);
@@ -74,8 +95,11 @@ const deleteProduct = async (req, res) => {
 /*--------------------------------------------Get a single product by ID---------------------------------------*/
 
 const getProduct = async (req, res) => {
-    const { productId } = req.params;
 
+    const { productId } = req.params;
+    if (!isValidObjectId(productId)) {
+        return res.status(400).json(new ApiError(400, null, "Invalid product ID"));
+    }
     try {
         const product = await Product.findById(productId);
         if (!product) {
@@ -91,9 +115,25 @@ const getProduct = async (req, res) => {
 /*--------------------------------------------Get all products---------------------------------------*/
 
 const getAllProducts = async (req, res) => {
+    const { page = 1, limit = 10 } = req.query;
+    console.log("checkk",page)
+    const pageNumber = parseInt(page);
+    const limitNumber = parseInt(limit);
+    const skip = (pageNumber - 1) * limitNumber;
     try {
-        const products = await Product.find();
-        return res.status(200).json(new ApiResponse(200, products, "Products retrieved successfully!"));
+
+        const products = await Product.find().skip(skip).limit(limitNumber);
+        const totalProducts = await Product.countDocuments()
+        return res.status(200).json(new ApiResponse(200,
+            {
+                result: products,
+                pagination: {
+                    currentPage: pageNumber,
+                    totalPages: Math.ceil(totalProducts / limitNumber),
+                    totalItems: totalProducts,
+                    itemsPerPage: limitNumber,
+                },
+            }, "Products retrieved successfully!"));
     } catch (error) {
         console.error(error);
         return res.status(500).json(new ApiResponse(500, "Something went wrong", [error.message]));
